add counter as array of indices to Unit
In order to use i=j for-loop optimization I've use counter field as
array of indices (to prevent race conditions). The algorithm run
slower, because sees is more complicated, and counter update is more
complicated too.

But if you think in terms of real life use cases of
this program, then this might be interesting, because often you will
need to get Units which this Unit is seeing, not just their count.